AL-28 modified tabs to use allowedContent from selected assignment, fixed a few styling issues
