+++ portal/src/utils/formatters.js
@@ -6,3 +6,11 @@ export function formatDate(upcomingDate) {
   const formattedDate = `${month} ${day}`;
   return formattedDate.toUpperCase();
 }
+
+export function formatDateMDY(dateStr) {
+  const date = new Date(dateStr);
+  const month = String(date.getMonth() + 1);
+  const day = String(date.getDate());
+  const year = date.getFullYear().toString().slice(-2);
+  return `${month}/${day}/${year}`;
+}
